Remove strategy structs for simplification
